fix: editabletext now uses his own state
