+++ packages/containers/src/SubHeaderBar/SubHeaderBar.test.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import { Map } from 'immutable';
 import Container, { DEFAULT_STATE, DISPLAY_NAME } from './SubHeaderBar.container';
 import Connect from './SubHeaderBar.connect';
-import { getComponentState, getEditMode } from './SubHeaderBar.selectors';
+import { getComponentState } from './SubHeaderBar.selectors';
 
 describe('Connect', () => {
 	it('should connect SubHeaderBar', () => {
@@ -16,148 +16,6 @@ describe('SubHeaderBar container', () => {
 	it('should render', () => {
 		const wrapper = shallow(<Container />);
 		expect(wrapper.getElement()).toMatchSnapshot();
-	});
-	it('should setState when submit event trigger', () => {
-		const event = {};
-		let state;
-		const props = {
-			state: Map({ editMode: true }),
-			setState: jest.fn(fn => (state = fn())),
-			onCancel: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('submit', event);
-		expect(props.setState).toHaveBeenCalled();
-		expect(state.editMode).toEqual(false);
-	});
-	it('should call onSubmit when submit event trigger', () => {
-		const event = {};
-		const data = { value: 'submitValue', props: {} };
-		const props = {
-			onSubmit: jest.fn(),
-			state: Map({ editMode: false }),
-			setState: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('submit', event, data);
-		expect(props.setState).toHaveBeenCalled();
-		expect(props.onSubmit).toHaveBeenCalledWith(event, data);
-	});
-	it('should call ActionCreatorSubmit when submit event trigger', () => {
-		const event = {};
-		const data = { value: 'submitValue', props: {} };
-		const props = {
-			actionCreatorSubmit: 'mySubmitActionCreator',
-			dispatchActionCreator: jest.fn(),
-			state: Map({ editMode: false }),
-			setState: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('submit', event, data);
-		expect(props.dispatchActionCreator).toHaveBeenCalledWith(props.actionCreatorSubmit, event, {
-			props,
-			data,
-		});
-	});
-	it('should setState when cancel event trigger', () => {
-		const event = {};
-		let state;
-		const props = {
-			state: Map({ editMode: true }),
-			setState: jest.fn(fn => (state = fn())),
-			onCancel: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('cancel', event);
-		expect(props.setState).toHaveBeenCalled();
-		expect(state.editMode).toEqual(false);
-	});
-	it('should call onCancel when cancel event trigger', () => {
-		const event = {};
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			onCancel: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('cancel', event);
-		expect(props.onCancel).toHaveBeenCalledWith(event);
-	});
-	it('should call actionCreatorCancel when cancel event trigger', () => {
-		const event = {};
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			actionCreatorCancel: 'myCancelActionCreator',
-			dispatchActionCreator: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('cancel', event);
-		expect(props.dispatchActionCreator).toHaveBeenCalledWith(props.actionCreatorCancel, event, {
-			props,
-		});
-	});
-	it('should call setState when edit event trigger', () => {
-		const event = {};
-		let state;
-		const props = {
-			state: Map({ editMode: true }),
-			setState: jest.fn(fn => (state = fn())),
-			onCancel: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('edit', event);
-		expect(props.setState).toHaveBeenCalled();
-		expect(state.editMode).toEqual(false);
-	});
-	it('should call onEdit when edit event trigger', () => {
-		const event = {};
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			onEdit: jest.fn(),
-		};
-		shallow(<Container {...props} />).simulate('edit', event);
-		expect(props.onEdit).toHaveBeenCalledWith(event);
-	});
-	it('should call onEdit when edit event trigger', () => {
-		// Given
-		const event = {};
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			dispatchActionCreator: jest.fn(),
-			actionCreatorEdit: 'myEditActionCreator',
-		};
-		// When
-		shallow(<Container {...props} />).simulate('edit', event);
-		// Then
-		expect(props.dispatchActionCreator).toHaveBeenCalledWith(props.actionCreatorEdit, event, {
-			props,
-		});
-	});
-	it('should call onChange when change event trigger', () => {
-		// Given
-		const event = { target: { value: 'my onChangeTitle' } };
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			onChange: jest.fn(),
-		};
-		// When
-		shallow(<Container {...props} />).simulate('change', event);
-		// Then
-		expect(props.onChange).toHaveBeenCalledWith(event, event.target.value);
-	});
-	it('should call onChange when change event tigger', () => {
-		// Given
-		const event = { target: { value: 'my onChangeTitle' } };
-		const props = {
-			setState: jest.fn(),
-			state: Map({ editMode: false }),
-			dispatchActionCreator: jest.fn(),
-			actionCreatorChange: 'myChangeActionCreator',
-		};
-		// When
-		shallow(<Container {...props} />).simulate('change', event);
-		// Then
-		expect(props.dispatchActionCreator).toHaveBeenCalledWith(props.actionCreatorChange, event, {
-			props,
-			value: event.target.value,
-		});
 	});
 	it('should call onGoBack event when goBack event trigger', () => {
 		// Given
@@ -188,7 +46,7 @@ describe('SubHeaderBar container', () => {
 
 describe('SubHeaderBar selectors', () => {
 	let mockState;
-	const componentState = Map({ editMode: true });
+	const componentState = Map({});
 	beforeEach(() => {
 		mockState = {
 			cmf: {
@@ -201,11 +59,5 @@ describe('SubHeaderBar selectors', () => {
 	});
 	it('should return the default state', () => {
 		expect(getComponentState(mockState, 'wrongComponentId')).toEqual(DEFAULT_STATE);
-	});
-	it('should return the editMode', () => {
-		expect(getEditMode(mockState, 'mySubHeaderBar')).toEqual(true);
-	});
-	it('should return the editModet', () => {
-		expect(getEditMode(mockState, 'wrongComponentId')).toEqual(false);
 	});
 });
